requirements.txt
imutils==0.5.4
numpy==1.26.3
pillow==10.2.0
pypng==0.20220715.0
PySimpleGUI==4.60.5
pyzbar==0.1.9
qrcode==7.4.2
typing_extensions==4.9.0
zbar-py==1.0.4
